TST: Reorganize analysis testing functions into different scripts
Reorganize analysis testing functions into different scripts: split the
existing `test_analysis.py` script contents into two scripts each
containing testing corresponding to the `analysis.motion` and
`analysis.measure_agreement` modules.

Follows the approach already in place for the other `analysis` module
(`filtering`), and improves compartmentalization.
