+++ test/test_measure_agreement.py
@@ -20,7 +20,7 @@
 #
 #     https://www.nipreps.org/community/licensing/
 #
-"""Unit tests exercising the analysis."""
+"""Unit tests exercising measure agreement analysis."""
 
 import numpy as np
 import pytest
@@ -31,7 +31,6 @@ from nifreeze.analysis.measure_agreement import (
     compute_z_score,
     identify_bland_altman_salient_data,
 )
-from nifreeze.analysis.motion import identify_spikes
 
 
 def test_compute_z_score():
@@ -142,21 +141,3 @@ def test_identify_bland_altman_salient_data():
 
     assert len(salient_data[BASalientEntity.RIGHT_INDICES.value]) == top_n
     assert len(salient_data[BASalientEntity.RIGHT_MASK.value]) == len(_data1)
-
-
-def test_identify_spikes():
-    rng = np.random.default_rng(1234)
-
-    n_samples = 450
-
-    fd = rng.normal(0, 5, n_samples)
-    threshold = 2.0
-
-    expected_indices = np.asarray([5, 57, 85, 100, 127, 180, 191, 202, 335, 393, 409])
-    expected_mask = np.zeros(n_samples, dtype=bool)
-    expected_mask[expected_indices] = True
-
-    obtained_indices, obtained_mask = identify_spikes(fd, threshold=threshold)
-
-    assert np.array_equal(obtained_indices, expected_indices)
-    assert np.array_equal(obtained_mask, expected_mask)
